+++ registry/api/v2/routes.go
@@ -13,7 +13,39 @@ const (
 	RouteNameBlobUpload      = "blob-upload"
 	RouteNameBlobUploadChunk = "blob-upload-chunk"
 	RouteNameCatalog         = "catalog"
+
+	RoutePathBase            = "/v2/"
+	RoutePathManifest        = "/v2/{name}/manifests/{reference}"
+	RoutePathTags            = "/v2/{name}/tags/list"
+	RoutePathTag             = "/v2/{name}/tags/reference/{tag}"
+	RoutePathBlob            = "/v2/{name}/blobs/{digest}"
+	RoutePathBlobUpload      = "/v2/{name}/blobs/uploads/"
+	RoutePathBlobUploadChunk = "/v2/{name}/blobs/uploads/{uuid}"
+	RoutePathCatalog         = "/v2/_catalog"
 )
+
+func RoutePath(routeName string) string {
+	switch routeName {
+	case RouteNameBase:
+		return RoutePathBase
+	case RouteNameManifest:
+		return RoutePathManifest
+	case RouteNameTags:
+		return RoutePathTags
+	case RouteNameTag:
+		return RoutePathTag
+	case RouteNameBlob:
+		return RoutePathBlob
+	case RouteNameBlobUpload:
+		return RoutePathBlobUpload
+	case RouteNameBlobUploadChunk:
+		return RoutePathBlobUploadChunk
+	case RouteNameCatalog:
+		return RoutePathCatalog
+	default:
+		return ""
+	}
+}
 
 // Router builds a gorilla router with named routes for the various API
 // methods. This can be used directly by both server implementations and