+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ## Added
 - registry: Experimental PostgreSQL metadata database (incomplete, in progress)
+- registry: Use GitLab LabKit for HTTP metrics collection
 
 ## [v2.9.1-gitlab] - 2020-05-05
 ## Added